Trying to fix sort to properly sort more than about 2 billion suffixes, by switching to 64 bit ints.

+++ libsuffixtools/Concurrency/MkqsThread.h
@@ -14,10 +14,10 @@
 template<typename T>
 struct MkqsJob
 {
-    MkqsJob(T* p, int num, int d) : pData(p), n(num), depth(d) {}
+    MkqsJob(T* p, size_t num, size_t d) : pData(p), n(num), depth(d) {}
     T* pData;
-    int n;
-    int depth;
+    size_t n;
+    size_t depth;
 };
 
 //
